#3042 rename + relocate Utils file
