+++ wls-gui-wahllokalsystem/src/composables/appUtils.ts
